requirements-test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file=requirements-test.txt requirements-test.in
#
apipkg==1.5               # via execnet
appdirs==1.4.3            # via black
aspy.yaml==1.3.0          # via pre-commit
asynctest==0.13.0
atomicwrites==1.3.0       # via pytest
attrs==19.3.0             # via black, jsonschema, pytest
aws-sam-translator==1.15.1  # via cfn-lint
aws-xray-sdk==2.4.2       # via moto
bandit==1.6.2
black==19.10b0
bleach==3.1.0             # via readme-renderer
boto3==1.10.16            # via aws-sam-translator, moto
boto==2.49.0              # via moto
botocore==1.13.16         # via aws-xray-sdk, boto3, moto, s3transfer
certifi==2019.9.11        # via requests
cffi==1.13.2              # via cryptography
cfgv==2.0.1               # via pre-commit
cfn-lint==0.25.0          # via moto
chardet==3.0.4            # via requests
click==7.0                # via black
coverage==4.5.4
cryptography==2.8         # via moto, sshpubkeys
docker==4.1.0             # via moto
docutils==0.15.2          # via botocore, readme-renderer
ecdsa==0.14.1             # via python-jose, sshpubkeys
entrypoints==0.3          # via flake8
execnet==1.7.1            # via pytest-xdist
filelock==3.0.12          # via tox
flake8-docstrings==1.5.0
flake8-import-order==0.18.1
flake8==3.7.9
future==0.18.2            # via aws-xray-sdk, python-jose
gitdb2==2.0.6             # via gitpython
gitpython==3.0.4          # via bandit
identify==1.4.7           # via pre-commit
idna==2.8                 # via moto, requests
importlib-metadata==0.23  # via jsonschema, pluggy, pre-commit, pytest, tox
jinja2==2.10.3            # via moto
jmespath==0.9.4           # via boto3, botocore
jsondiff==1.1.2           # via moto
jsonpatch==1.24           # via cfn-lint
jsonpickle==1.2           # via aws-xray-sdk
jsonpointer==2.0          # via jsonpatch
jsonschema==3.1.1         # via aws-sam-translator, cfn-lint
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via flake8
mock==3.0.5
mockredispy==2.9.3
more-itertools==7.2.0     # via pytest, zipp
moto==1.3.14.dev522
mypy-extensions==0.4.3    # via mypy
mypy==0.740
nodeenv==1.3.3            # via pre-commit
packaging==19.2           # via pytest, tox
pathspec==0.6.0           # via black
pbr==5.4.3                # via stevedore
pillow==6.2.1
pluggy==0.13.0            # via pytest, tox
pre-commit==1.20.0
py==1.8.0                 # via pytest, tox
pyasn1==0.4.7             # via rsa
pycodestyle==2.5.0        # via flake8, flake8-import-order
pycparser==2.19           # via cffi
pydocstyle==4.0.1         # via flake8-docstrings
pyflakes==2.1.1           # via flake8
pygments==2.4.2           # via readme-renderer
pyparsing==2.4.5          # via packaging
pyrsistent==0.15.5        # via jsonschema
pytest-asyncio==0.10.0
pytest-cov==2.8.1
pytest-forked==1.1.3      # via pytest-xdist
pytest-timeout==1.3.3
pytest-tornado==0.8.0
pytest-xdist==1.30.0
pytest==5.2.2
python-dateutil==2.8.0    # via botocore, moto
python-jose==3.0.1        # via moto
pytz==2019.3              # via moto
pyyaml==5.1.2             # via aspy.yaml, bandit, cfn-lint, moto, pre-commit
readme-renderer==24.0
regex==2019.11.1          # via black
requests==2.22.0          # via docker, moto, responses
responses==0.10.6         # via moto
rsa==4.0                  # via python-jose
s3transfer==0.2.1         # via boto3
selenium==3.141.0
six==1.13.0               # via aws-sam-translator, bandit, bleach, cfgv, cfn-lint, cryptography, docker, ecdsa, jsonschema, mock, moto, packaging, pre-commit, pyrsistent, pytest-xdist, python-dateutil, python-jose, readme-renderer, responses, stevedore, tox, websocket-client
smmap2==2.0.5             # via gitdb2
snowballstemmer==2.0.0    # via pydocstyle
sshpubkeys==3.1.0         # via moto
stevedore==1.31.0         # via bandit
toml==0.10.0              # via black, pre-commit, tox
tornado==6.0.3
tox==3.14.1
typed-ast==1.4.0          # via black, mypy
typing-extensions==3.7.4.1  # via mypy
urllib3==1.25.7           # via botocore, requests, selenium
virtualenv==16.7.7        # via pre-commit, tox
wcwidth==0.1.7            # via pytest
webencodings==0.5.1       # via bleach
websocket-client==0.56.0  # via docker
werkzeug==0.16.0          # via moto
wrapt==1.11.2             # via aws-xray-sdk
xmltodict==0.12.0         # via moto
zipp==0.6.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.6.0        # via flake8-import-order, jsonschema
